resume_classifier/requirements.txt
Flask==2.3.3
openai==1.3.0
PyPDF2==3.0.1
python-docx==0.8.11
Werkzeug==2.3.7
SQLAlchemy>=2.0.25
Flask-SQLAlchemy>=3.1.0
python-dotenv==1.0.0
requests==2.31.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0 